Create README.md
NeonBlast is a Neon Theme for Emulationstation Which so far has 9 system views and 7 gamelist views a couple of which are a work in progress. Big thanks to Dwayne Hurst for allowing me to repurpose some of his assets from Arcadeparadise for menu 9 and gamelist cabs view. I also want to thank EvilDinDon as i used 1 of his animations. To customise your theme a couple of the views utilize your fanart other customisations can include simply copying, renaming and changing any frames from their folders. Also includes 3 styles of wallpapers and animation on or off. Enjoy!!!
|                                    |                                    |                                    
| :--------------------------------: | :--------------------------------: | 
| ![](https://imgur.com/NVhwBuj.jpg?raw=true) | ![](https://imgur.com/DIQXQT3.jpg?raw=true) | ![](https://imgur.com/dLnicB2.jpg?raw=true) |
| ![](https://imgur.com/7ZvTDUM.jpg?raw=true) | ![](https://imgur.com/EZxRRs1.jpg?raw=true) | ![](https://imgur.com/dSfP3KT.jpg?raw=true) |